+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [4.13.0] - 2020-06-27
 ### `jsonschema-generator`
 #### Added
 - Possibility to configure `SchemaDefinitionNamingStrategy` via `configBuilder.forTypesInGeneral().withDefinitionNamingStrategy()`
@@ -24,6 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### `jsonschema-module-swagger-2`
 #### Added
 - Initial implementation of `Swagger2Module` for deriving schema attributes from OpenAPI `@Schema` annotations.
+
+### `jsonschema-maven-plugin`
+#### Added
+- Support for new `jsonschema-module-swagger-2`
 
 ## [4.12.2] - 2020-06-10
 ### `jsonschema-generator`
@@ -414,7 +418,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Indicate a number's "exclusiveMaximum" according to `@DecimalMax` or `@Negative`
 
 
-[Unreleased]: https://github.com/victools/jsonschema-generator/compare/v4.12.2...HEAD
+[4.13.0]: https://github.com/victools/jsonschema-generator/compare/v4.12.2...v4.13.0
 [4.12.2]: https://github.com/victools/jsonschema-generator/compare/v4.12.1...v4.12.2
 [4.12.1]: https://github.com/victools/jsonschema-generator/compare/v4.12.0...v4.12.1
 [4.12.0]: https://github.com/victools/jsonschema-generator/compare/v4.11.1...v4.12.0